test: move action tests to separate files
this reflects the new file structure under src/router directory
and makes everything more clear.

Also use common helper to get test data and test plugins
