+++ core/test/unit/src/actions.ts
@@ -6,122 +6,41 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import {
-  ProviderHandlers,
-  getModuleHandlerDescriptions,
-  getProviderActionDescriptions,
-  createGardenPlugin,
-  ActionHandler,
-  ModuleActionHandler,
-} from "../../../src/plugin/plugin"
+import { createGardenPlugin, ActionHandler, ModuleActionHandler } from "../../../src/plugin/plugin"
 import { ServiceState } from "../../../src/types/service"
-import { expectError, makeTestGardenA, stubRouterAction, projectRootA, TestGarden, makeTestGarden } from "../../helpers"
+import { expectError, makeTestGardenA, projectRootA, TestGarden, makeTestGarden } from "../../helpers"
 import { ActionRouter } from "../../../src/router/router"
 import { LogEntry } from "../../../src/logger/log-entry"
-import { GardenModule } from "../../../src/types/module"
-import { ServiceLogEntry } from "../../../src/types/service"
-import Stream from "ts-stream"
 import { expect } from "chai"
-import { cloneDeep, omit } from "lodash"
-import { CustomObjectSchema, joi } from "../../../src/config/common"
-import { validateSchema } from "../../../src/config/validation"
+import { joi } from "../../../src/config/common"
 import { ProjectConfig, defaultNamespace } from "../../../src/config/project"
 import { DEFAULT_API_VERSION } from "../../../src/constants"
 import { defaultProvider, providerFromConfig } from "../../../src/config/provider"
 import { defaultDotIgnoreFile } from "../../../src/util/fs"
-import stripAnsi from "strip-ansi"
-import { emptyDir, pathExists, ensureFile, readFile } from "fs-extra"
-import { join } from "path"
 import { DashboardPage } from "../../../src/plugin/handlers/provider/getDashboardPage"
 import { ConfigGraph } from "../../../src/graph/config-graph"
-import { BuildActionConfig, ResolvedBuildAction } from "../../../src/actions/build"
-import {
-  execBuildActionSchema,
-  execDeployActionSchema,
-  execRunActionSchema,
-  execTestActionSchema,
-} from "../../../src/plugins/exec/config"
-import { convertModules } from "../../../src/resolve-module"
-import { actionFromConfig } from "../../../src/graph/actions"
-import { TestAction, TestActionConfig } from "../../../src/actions/test"
-import { TestConfig } from "../../../src/config/test"
-import { findByName } from "../../../src/util/util"
 import { ResolvedRunAction } from "../../../src/actions/run"
-import { ResolvedDeployAction } from "../../../src/actions/deploy"
-import { BaseActionConfig } from "../../../src/actions/types"
-import { BaseRuntimeActionConfig } from "../../../src/actions/base"
-
-const now = new Date()
-const returnWrongOutputsCfgKey = "returnWrong"
+import { getRouterTestData } from "./router/_helpers"
 
 describe("ActionRouter", () => {
   let garden: TestGarden
   let graph: ConfigGraph
   let log: LogEntry
   let actionRouter: ActionRouter
-  let module: GardenModule
-  let resolvedBuildAction: ResolvedBuildAction
-  let resolvedDeployAction: ResolvedDeployAction
   let resolvedRunAction: ResolvedRunAction
 
-  const projectConfig: ProjectConfig = {
-    apiVersion: DEFAULT_API_VERSION,
-    kind: "Project",
-    name: "test",
-    path: projectRootA,
-    defaultEnvironment: "default",
-    dotIgnoreFile: defaultDotIgnoreFile,
-    environments: [{ name: "default", defaultNamespace, variables: {} }],
-    providers: [{ name: "base" }, { name: "test-plugin" }, { name: "test-plugin-b" }],
-    variables: {},
-  }
-
   before(async () => {
-    garden = await makeTestGarden(projectRootA, {
-      plugins: [basePlugin, testPlugin, testPluginB],
-      config: projectConfig,
-    })
-    projectConfig.path = garden.projectRoot
-    log = garden.log
-    actionRouter = await garden.getActionRouter()
-    graph = await garden.getConfigGraph({ log: garden.log, emit: false })
-    module = graph.getModule("module-a")
-    const buildAction = graph.getBuild("module-a")
-    resolvedBuildAction = await garden.resolveAction({
-      action: buildAction,
-      log: garden.log,
-      graph: await garden.getConfigGraph({ log: garden.log, emit: false }),
-    })
-    const deployAction = graph.getDeploy("service-a")
-    resolvedDeployAction = await garden.resolveAction({
-      action: deployAction,
-      log: garden.log,
-      graph: await garden.getConfigGraph({ log: garden.log, emit: false }),
-    })
-    const runAction = graph.getRun("task-a")
-    resolvedRunAction = await garden.resolveAction({
-      action: runAction,
-      log: garden.log,
-      graph: await garden.getConfigGraph({ log: garden.log, emit: false }),
-    })
+    const data = await getRouterTestData()
+    garden = data.garden
+    graph = data.graph
+    log = data.log
+    actionRouter = data.actionRouter
+    resolvedRunAction = data.resolvedRunAction
   })
 
   after(async () => {
     await garden.close()
   })
-
-  async function getResolvedAction(testConfig: TestActionConfig<string, any>) {
-    const action = (await actionFromConfig({
-      garden,
-      // rebuild config graph because the module config has been changed
-      graph: await garden.getConfigGraph({ emit: false, log: garden.log }),
-      config: testConfig,
-      log: garden.log,
-      configsByKey: {},
-      router: await garden.getActionRouter(),
-    })) as TestAction
-    return await garden.resolveAction<TestAction>({ action, log: garden.log })
-  }
 
   // Note: The test plugins below implicitly validate input params for each of the tests
   describe("environment actions", () => {
@@ -301,594 +220,11 @@ describe("ActionRouter", () => {
         ])
       })
     })
-
-    describe("build.getStatus", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.build.getStatus({ log, action: resolvedBuildAction, graph })
-        expect(result).to.eql({
-          detail: {},
-          outputs: {
-            foo: "bar",
-          },
-          state: "ready",
-        })
-      })
-
-      it("should emit a buildStatus event", async () => {
-        garden.events.eventLog = []
-        await actionRouter.build.getStatus({ log, action: resolvedBuildAction, graph })
-        const event = garden.events.eventLog[0]
-        expect(event).to.exist
-        expect(event.name).to.eql("buildStatus")
-        expect(event.payload.moduleName).to.eql("module-a")
-        expect(event.payload.moduleVersion).to.eql(module.version.versionString)
-        expect(event.payload.actionUid).to.be.undefined
-        expect(event.payload.status.state).to.eql("fetched")
-      })
-    })
-
-    describe("build", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.build.build({ log, action: resolvedBuildAction, graph })
-        expect(result).to.eql({
-          detail: {},
-          outputs: {
-            foo: "bar",
-          },
-          state: "ready",
-        })
-      })
-
-      it("should emit buildStatus events", async () => {
-        garden.events.eventLog = []
-        await actionRouter.build.build({ log, action: resolvedBuildAction, graph })
-        const event1 = garden.events.eventLog[0]
-        const event2 = garden.events.eventLog[1]
-        const moduleVersion = module.version.versionString
-        expect(event1).to.exist
-        expect(event1.name).to.eql("buildStatus")
-        expect(event1.payload.moduleName).to.eql("module-a")
-        expect(event1.payload.moduleVersion).to.eql(moduleVersion)
-        expect(event1.payload.status.state).to.eql("building")
-        expect(event1.payload.actionUid).to.be.ok
-        expect(event2).to.exist
-        expect(event2.name).to.eql("buildStatus")
-        expect(event2.payload.moduleName).to.eql("module-a")
-        expect(event2.payload.moduleVersion).to.eql(moduleVersion)
-        expect(event2.payload.status.state).to.eql("built")
-        expect(event2.payload.actionUid).to.eql(event1.payload.actionUid)
-      })
-    })
-
-    describe("build.run", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const command = ["npm", "run"]
-        const result = await actionRouter.build.run({
-          log,
-          action: await garden.executeAction({
-            action: resolvedBuildAction,
-            log: garden.log,
-            graph: await garden.getConfigGraph({ log: garden.log, emit: false }),
-          }),
-          args: command,
-          interactive: true,
-          graph,
-        })
-        expect(result).to.eql({
-          moduleName: module.name,
-          command,
-          completedAt: now,
-          log: "bla bla",
-          success: true,
-          startedAt: now,
-          version: resolvedBuildAction.versionString(),
-        })
-      })
-    })
-
-    describe("test.run", () => {
-      const actionConfig: TestActionConfig = {
-        name: "test",
-        type: "test",
-        internal: { basePath: "test" },
-        kind: "Test",
-        dependencies: [],
-        disabled: false,
-        timeout: 1234,
-        spec: {},
-      }
-
-      it("should correctly call the corresponding plugin handler", async () => {
-        const action = await getResolvedAction(actionConfig)
-        const result = await actionRouter.test.run({
-          log,
-          action,
-          interactive: true,
-          graph,
-          silent: false,
-        })
-        expect(result.outputs).to.eql({})
-        expect(result.detail?.log).to.eql("bla bla")
-        expect(result.state).to.eql("ready")
-      })
-
-      it("should emit testStatus events", async () => {
-        const action = await getResolvedAction(actionConfig)
-        garden.events.eventLog = []
-        await actionRouter.test.run({
-          log,
-          action,
-          interactive: true,
-          graph,
-          silent: false,
-        })
-        const testVersion = action.versionString()
-        const event1 = garden.events.eventLog[0]
-        const event2 = garden.events.eventLog[1]
-        expect(event1).to.exist
-        expect(event1.name).to.eql("testStatus")
-        expect(event1.payload.testName).to.eql("test")
-        expect(event1.payload.actionName).to.eql("test")
-        expect(event1.payload.testVersion).to.eql(testVersion)
-        expect(event1.payload.actionUid).to.be.ok
-        expect(event1.payload.status.state).to.eql("running")
-        expect(event2).to.exist
-        expect(event2.name).to.eql("testStatus")
-        expect(event2.payload.testName).to.eql("test")
-        expect(event2.payload.actionName).to.eql("test")
-        expect(event2.payload.testVersion).to.eql(testVersion)
-        expect(event2.payload.actionUid).to.eql(event1.payload.actionUid)
-        expect(event2.payload.status.state).to.eql("succeeded")
-      })
-
-      it("should copy artifacts exported by the handler to the artifacts directory", async () => {
-        await emptyDir(garden.artifactsPath)
-
-        const testConfig = {
-          ...actionConfig,
-          spec: {
-            artifacts: [
-              {
-                source: "some-file.txt",
-              },
-              {
-                source: "some-dir/some-file.txt",
-                target: "some-dir/some-file.txt",
-              },
-            ],
-          },
-        }
-
-        const action = await getResolvedAction(testConfig)
-
-        await actionRouter.test.run({
-          log,
-          action,
-          interactive: true,
-          graph,
-          silent: false,
-        })
-
-        const targetPaths = testConfig.spec.artifacts.map((spec) => join(garden.artifactsPath, spec.source)).sort()
-
-        for (const path of targetPaths) {
-          expect(await pathExists(path)).to.be.true
-        }
-
-        const metadataKey = `test.test.${action.versionString()}`
-        const metadataFilename = `.metadata.${metadataKey}.json`
-        const metadataPath = join(garden.artifactsPath, metadataFilename)
-        expect(await pathExists(metadataPath)).to.be.true
-
-        const metadata = JSON.parse((await readFile(metadataPath)).toString())
-        expect(metadata).to.eql({
-          key: metadataKey,
-          files: targetPaths,
-        })
-      })
-    })
-
-    describe("test.getResult", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const action = await garden.resolveAction({ action: graph.getTest("module-a-unit"), log, graph })
-
-        const result = await actionRouter.test.getResult({
-          log,
-          action,
-          graph,
-        })
-
-        expect(result.outputs).to.eql({})
-        expect(result.detail?.log).to.eql("bla bla")
-        expect(result.state).to.eql("ready")
-      })
-    })
-
-    it("should emit a testStatus event", async () => {
-      const action = await garden.resolveAction({ action: graph.getTest("module-a-unit"), log, graph })
-      garden.events.eventLog = []
-
-      await actionRouter.test.getResult({
-        log,
-        action,
-        graph,
-      })
-      const event = garden.events.eventLog[0]
-
-      expect(event).to.exist
-      expect(event.name).to.eql("testStatus")
-      expect(event.payload.testName).to.eql("module-a-unit")
-      expect(event.payload.moduleVersion).to.eql(module.version.versionString)
-      expect(event.payload.testVersion).to.eql(action.versionString())
-      expect(event.payload.actionUid).to.be.undefined
-      expect(event.payload.status.state).to.eql("succeeded")
-    })
   })
 
-  describe("deploy actions", () => {
-    describe("deploy.getStatus", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.deploy.getStatus({
-          log,
-          action: resolvedDeployAction,
-          graph,
-          devMode: false,
+  describe("getHandler", () => {
+    const path = projectRootA
 
-          localMode: false,
-        })
-        expect(result).to.eql({
-          detail: { forwardablePorts: [], state: "ready", outputs: {}, detail: {} },
-          outputs: { base: "ok", foo: "ok" },
-          state: "ready",
-        })
-      })
-
-      it("should emit a serviceStatus event", async () => {
-        garden.events.eventLog = []
-        await actionRouter.deploy.getStatus({
-          log,
-          action: resolvedDeployAction,
-          graph,
-          devMode: false,
-          localMode: false,
-        })
-        const event = garden.events.eventLog[0]
-        expect(event).to.exist
-        expect(event.name).to.eql("serviceStatus")
-        expect(event.payload.serviceName).to.eql("service-a")
-        expect(event.payload.actionVersion).to.eql(resolvedDeployAction.versionString())
-        expect(event.payload.serviceVersion).to.eql(resolvedDeployAction.versionString())
-        expect(event.payload.actionUid).to.be.undefined
-        expect(event.payload.status.state).to.eql("ready")
-      })
-
-      it("should throw if the outputs don't match the service outputs schema of the plugin", async () => {
-        const action = cloneDeep(resolvedDeployAction)
-        action._config[returnWrongOutputsCfgKey] = true
-        await expectError(
-          () =>
-            actionRouter.deploy.getStatus({
-              log,
-              action,
-              graph,
-              devMode: false,
-              localMode: false,
-            }),
-          (err) =>
-            expect(stripAnsi(err.message)).to.include(
-              "Error validating outputs from deploy 'service-a': key .foo must be a string"
-            )
-        )
-      })
-    })
-
-    describe("deploy.deploy", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.deploy.deploy({
-          log,
-          action: resolvedDeployAction,
-          graph,
-          force: true,
-          devMode: false,
-          localMode: false,
-        })
-        expect(result).to.eql({
-          detail: { forwardablePorts: [], state: "ready", outputs: {}, detail: {} },
-          outputs: { base: "ok", foo: "ok" },
-          state: "ready",
-        })
-      })
-
-      it("should emit serviceStatus events", async () => {
-        garden.events.eventLog = []
-        await actionRouter.deploy.deploy({
-          log,
-          action: resolvedDeployAction,
-          graph,
-          force: true,
-          devMode: false,
-          localMode: false,
-        })
-        const moduleVersion = resolvedDeployAction.moduleVersion().versionString
-        const event1 = garden.events.eventLog[0]
-        const event2 = garden.events.eventLog[1]
-        expect(event1).to.exist
-        expect(event1.name).to.eql("serviceStatus")
-        expect(event1.payload.serviceName).to.eql("service-a")
-        expect(event1.payload.moduleName).to.eql("module-a")
-        expect(event1.payload.moduleVersion).to.eql(moduleVersion)
-        expect(event1.payload.serviceVersion).to.eql(resolvedDeployAction.versionString())
-        expect(event1.payload.actionUid).to.be.ok
-        expect(event1.payload.status.state).to.eql("deploying")
-        expect(event2).to.exist
-        expect(event2.name).to.eql("serviceStatus")
-        expect(event2.payload.serviceName).to.eql("service-a")
-        expect(event2.payload.moduleName).to.eql("module-a")
-        expect(event2.payload.moduleVersion).to.eql(moduleVersion)
-        expect(event2.payload.serviceVersion).to.eql(resolvedDeployAction.versionString())
-        expect(event2.payload.actionUid).to.eql(event2.payload.actionUid)
-        expect(event2.payload.status.state).to.eql("ready")
-      })
-
-      it("should throw if the outputs don't match the service outputs schema of the plugin", async () => {
-        const action = cloneDeep(resolvedDeployAction)
-        action._config[returnWrongOutputsCfgKey] = true
-        await expectError(
-          () =>
-            actionRouter.deploy.deploy({
-              log,
-              action,
-              graph,
-              force: true,
-              devMode: false,
-              localMode: false,
-            }),
-          (err) =>
-            expect(stripAnsi(err.message)).to.include(
-              "Error validating outputs from deploy 'service-a': key .foo must be a string"
-            )
-        )
-      })
-    })
-
-    describe("deploy.delete", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.deploy.delete({ log, action: resolvedDeployAction, graph })
-        expect(result).to.eql({
-          forwardablePorts: [],
-          state: "ready",
-          detail: {
-            detail: {},
-            state: "ready",
-          },
-          outputs: {},
-        })
-      })
-    })
-
-    describe("deploy.exec", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const executedAction = await garden.executeAction({ action: resolvedDeployAction, log, graph })
-        const result = await actionRouter.deploy.exec({
-          log,
-          action: executedAction,
-          graph,
-          command: ["foo"],
-          interactive: false,
-        })
-        expect(result).to.eql({ code: 0, output: "bla bla" })
-      })
-    })
-
-    describe("deploy.getLogs", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const stream = new Stream<ServiceLogEntry>()
-        const result = await actionRouter.deploy.getLogs({
-          log,
-          action: resolvedDeployAction,
-          graph,
-          stream,
-          follow: false,
-          tail: -1,
-        })
-        expect(result).to.eql({})
-      })
-    })
-
-    describe("deploy.run", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.deploy.run({
-          log,
-          action: resolvedDeployAction,
-          interactive: true,
-          graph,
-        })
-        expect(result).to.eql({
-          moduleName: resolvedDeployAction.name,
-          command: ["foo"],
-          completedAt: now,
-          log: "bla bla",
-          success: true,
-          startedAt: now,
-          version: resolvedDeployAction.versionString(),
-        })
-      })
-    })
-  })
-
-  describe("task actions", () => {
-    let taskResult
-
-    before(() => {
-      taskResult = {
-        detail: {
-          command: ["foo"],
-          completedAt: now,
-          log: "bla bla",
-          moduleName: "task-a",
-          startedAt: now,
-          success: true,
-          taskName: "task-a",
-          version: resolvedRunAction.versionString(),
-        },
-        outputs: {
-          base: "ok",
-          foo: "ok",
-        },
-        state: "ready",
-      }
-    })
-
-    describe("run.getResult", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.run.getResult({
-          log,
-          action: resolvedRunAction,
-          graph,
-        })
-        expect(result).to.eql(taskResult)
-      })
-
-      it("should emit a taskStatus event", async () => {
-        garden.events.eventLog = []
-        await actionRouter.run.getResult({
-          log,
-          action: resolvedRunAction,
-          graph,
-        })
-        const event = garden.events.eventLog[0]
-        expect(event).to.exist
-        expect(event.name).to.eql("taskStatus")
-        expect(event.payload.taskName).to.eql("task-a")
-        expect(event.payload.moduleName).to.eql("module-a")
-        expect(event.payload.moduleVersion).to.eql(resolvedRunAction.moduleVersion().versionString)
-        expect(event.payload.taskVersion).to.eql(resolvedRunAction.versionString())
-        expect(event.payload.actionUid).to.be.undefined
-        expect(event.payload.status.state).to.eql("succeeded")
-      })
-
-      it("should throw if the outputs don't match the task outputs schema of the plugin", async () => {
-        const action = cloneDeep(resolvedRunAction)
-        action._config[returnWrongOutputsCfgKey] = true
-        await expectError(
-          () => actionRouter.run.getResult({ log, action, graph }),
-          (err) =>
-            expect(stripAnsi(err.message)).to.include(
-              "Error validating outputs from task 'task-a': key .foo must be a string"
-            )
-        )
-      })
-    })
-
-    describe("run.run", () => {
-      it("should correctly call the corresponding plugin handler", async () => {
-        const result = await actionRouter.run.run({
-          log,
-          action: resolvedRunAction,
-          interactive: true,
-          graph,
-        })
-        expect(result).to.eql(taskResult)
-      })
-
-      it("should emit taskStatus events", async () => {
-        garden.events.eventLog = []
-        await actionRouter.run.run({
-          log,
-          action: resolvedRunAction,
-          interactive: true,
-          graph,
-        })
-        const moduleVersion = resolvedRunAction.moduleVersion().versionString
-        const event1 = garden.events.eventLog[0]
-        const event2 = garden.events.eventLog[1]
-        expect(event1).to.exist
-        expect(event1.name).to.eql("taskStatus")
-        expect(event1.payload.taskName).to.eql("task-a")
-        expect(event1.payload.moduleName).to.eql("module-a")
-        expect(event1.payload.moduleVersion).to.eql(moduleVersion)
-        expect(event1.payload.taskVersion).to.eql(resolvedRunAction.versionString())
-        expect(event1.payload.actionUid).to.be.ok
-        expect(event1.payload.status.state).to.eql("running")
-        expect(event2).to.exist
-        expect(event2.name).to.eql("taskStatus")
-        expect(event2.payload.taskName).to.eql("task-a")
-        expect(event2.payload.moduleName).to.eql("module-a")
-        expect(event2.payload.moduleVersion).to.eql(moduleVersion)
-        expect(event2.payload.taskVersion).to.eql(resolvedRunAction.versionString())
-        expect(event2.payload.actionUid).to.eql(event1.payload.actionUid)
-        expect(event2.payload.status.state).to.eql("succeeded")
-      })
-
-      it("should throw if the outputs don't match the task outputs schema of the plugin", async () => {
-        const action = cloneDeep(resolvedRunAction)
-        action._config[returnWrongOutputsCfgKey] = true
-        await expectError(
-          () =>
-            actionRouter.run.run({
-              log,
-              action,
-              interactive: true,
-              graph,
-            }),
-          (err) =>
-            expect(stripAnsi(err.message)).to.include(
-              "Error validating outputs from Run 'task-a': key .foo must be a string"
-            )
-        )
-      })
-
-      it("should copy artifacts exported by the handler to the artifacts directory", async () => {
-        await emptyDir(garden.artifactsPath)
-
-        graph = await garden.getConfigGraph({ log: garden.log, emit: false })
-        const runActionTaskA = graph.getRun("task-a")
-
-        runActionTaskA.getConfig().spec.artifacts = [
-          {
-            source: "some-file.txt",
-          },
-          {
-            source: "some-dir/some-file.txt",
-            target: "some-dir/some-file.txt",
-          },
-        ]
-
-        await actionRouter.run.run({
-          log,
-          action: await garden.resolveAction({
-            action: runActionTaskA,
-            log: garden.log,
-            graph,
-          }),
-          interactive: true,
-          graph,
-        })
-
-        const targetPaths = runActionTaskA
-          .getConfig()
-          .spec.artifacts.map((spec) => join(garden.artifactsPath, spec.source))
-          .sort()
-
-        for (const path of targetPaths) {
-          expect(await pathExists(path)).to.be.true
-        }
-
-        const metadataKey = `run.task-a.${runActionTaskA.versionString()}`
-        const metadataFilename = `.metadata.${metadataKey}.json`
-        const metadataPath = join(garden.artifactsPath, metadataFilename)
-        expect(await pathExists(metadataPath)).to.be.true
-
-        const metadata = JSON.parse((await readFile(metadataPath)).toString())
-        expect(metadata).to.eql({
-          key: metadataKey,
-          files: targetPaths,
-        })
-      })
-    })
-  })
-
-  describe("getActionHandler", () => {
     it("should return the configured handler for specified action type and plugin name", async () => {
       const gardenA = await makeTestGardenA()
       const actionsA = await gardenA.getActionRouter()
@@ -912,10 +248,6 @@ describe("ActionRouter", () => {
         "plugin"
       )
     })
-  })
-
-  describe("getModuleActionHandler", () => {
-    const path = projectRootA
 
     it("should return default handler, if specified and no handler is available", async () => {
       const gardenA = await makeTestGardenA()
@@ -1786,499 +1118,3 @@ describe("ActionRouter", () => {
     })
   })
 })
-
-const baseOutputsSchema = () => joi.object().keys({ base: joi.string() })
-const testOutputSchema = () => baseOutputsSchema().keys({ foo: joi.string() })
-
-const outputsCfg = {
-  schema: joi.object().keys({
-    base: joi.string(),
-    foo: joi.string(),
-  }),
-}
-
-const basePlugin = createGardenPlugin({
-  name: "base",
-  createModuleTypes: [
-    {
-      name: "base",
-      docs: "bla bla bla",
-      moduleOutputsSchema: baseOutputsSchema(),
-      needsBuild: true,
-      handlers: {},
-    },
-  ],
-})
-
-const pluginActionDescriptions = getProviderActionDescriptions()
-const moduleActionDescriptions = getModuleHandlerDescriptions()
-
-const testPlugin = createGardenPlugin({
-  name: "test-plugin",
-  dependencies: [{ name: "base" }],
-
-  handlers: <ProviderHandlers>{
-    configureProvider: async (params) => {
-      validateParams(params, pluginActionDescriptions.configureProvider.paramsSchema)
-      return { config: params.config }
-    },
-
-    getEnvironmentStatus: async (params) => {
-      validateParams(params, pluginActionDescriptions.getEnvironmentStatus.paramsSchema)
-      return {
-        ready: false,
-        outputs: {},
-      }
-    },
-
-    augmentGraph: async (params) => {
-      validateParams(params, pluginActionDescriptions.augmentGraph.paramsSchema)
-
-      const actionName = "added-by-" + params.ctx.provider.name
-
-      return {
-        addDependencies: [
-          {
-            by: {
-              kind: "Deploy",
-              name: actionName,
-            },
-            on: {
-              kind: "Build",
-              name: actionName,
-            },
-          },
-        ],
-        addActions: [
-          {
-            kind: "Build",
-            name: actionName,
-            type: "container",
-            internal: {
-              basePath: ".",
-            },
-            spec: {},
-          },
-          {
-            kind: "Deploy",
-            name: actionName,
-            type: "container",
-            internal: {
-              basePath: ".",
-            },
-            spec: {},
-          },
-        ],
-      }
-    },
-
-    getDashboardPage: async (params) => {
-      validateParams(params, pluginActionDescriptions.getDashboardPage.paramsSchema)
-      return { url: "http://" + params.page.name }
-    },
-
-    getDebugInfo: async (params) => {
-      validateParams(params, pluginActionDescriptions.getDebugInfo.paramsSchema)
-      return { info: {} }
-    },
-
-    prepareEnvironment: async (params) => {
-      validateParams(params, pluginActionDescriptions.prepareEnvironment.paramsSchema)
-      return { status: { ready: true, outputs: {} } }
-    },
-
-    cleanupEnvironment: async (params) => {
-      validateParams(params, pluginActionDescriptions.cleanupEnvironment.paramsSchema)
-      return {}
-    },
-
-    getSecret: async (params) => {
-      validateParams(params, pluginActionDescriptions.getSecret.paramsSchema)
-      return { value: params.key }
-    },
-
-    setSecret: async (params) => {
-      validateParams(params, pluginActionDescriptions.setSecret.paramsSchema)
-      return {}
-    },
-
-    deleteSecret: async (params) => {
-      validateParams(params, pluginActionDescriptions.deleteSecret.paramsSchema)
-      return { found: true }
-    },
-  },
-
-  createModuleTypes: [
-    {
-      name: "test",
-      base: "base",
-      docs: "bla bla bla",
-      moduleOutputsSchema: testOutputSchema(),
-      schema: joi.object(),
-      needsBuild: true,
-      title: "Bla",
-
-      handlers: {
-        configure: async (params) => {
-          validateParams(params, moduleActionDescriptions.configure.paramsSchema)
-
-          const serviceConfigs = params.moduleConfig.spec.services.map((spec) => ({
-            name: spec.name,
-            dependencies: spec.dependencies || [],
-            disabled: false,
-
-            spec,
-          }))
-
-          const taskConfigs = (params.moduleConfig.spec.tasks || []).map((spec) => ({
-            name: spec.name,
-            dependencies: spec.dependencies || [],
-            disabled: false,
-            spec,
-          }))
-
-          const testConfigs = (params.moduleConfig.spec.tests || []).map((spec) => ({
-            name: spec.name,
-            dependencies: spec.dependencies || [],
-            disabled: false,
-            spec,
-          }))
-
-          return {
-            moduleConfig: {
-              ...params.moduleConfig,
-              serviceConfigs,
-              taskConfigs,
-              testConfigs,
-            },
-          }
-        },
-
-        convert: async (params) => {
-          validateParams(params, moduleActionDescriptions.convert.paramsSchema)
-
-          const {
-            module,
-            services,
-            tasks,
-            tests,
-            dummyBuild,
-            convertBuildDependency,
-            convertRuntimeDependencies,
-          } = params
-
-          const actions: (BuildActionConfig | BaseRuntimeActionConfig)[] = []
-
-          const buildAction: BuildActionConfig = {
-            kind: "Build",
-            type: "test",
-            name: module.name,
-
-            ...params.baseFields,
-            ...dummyBuild,
-
-            dependencies: module.build.dependencies.map(convertBuildDependency),
-
-            spec: {
-              command: module.spec.build?.command,
-              env: module.spec.env,
-            },
-          }
-
-          actions.push(buildAction)
-
-          for (const service of services) {
-            actions.push({
-              kind: "Deploy",
-              type: "test",
-              name: service.name,
-              ...params.baseFields,
-
-              disabled: service.disabled,
-              build: buildAction ? buildAction.name : undefined,
-              dependencies: convertRuntimeDependencies(service.spec.dependencies),
-
-              spec: {
-                ...omit(service.spec, ["name", "dependencies", "disabled"]),
-              },
-            })
-          }
-
-          for (const task of tasks) {
-            actions.push({
-              kind: "Run",
-              type: "test",
-              name: task.name,
-              ...params.baseFields,
-
-              disabled: task.disabled,
-              build: buildAction ? buildAction.name : undefined,
-              dependencies: convertRuntimeDependencies(task.spec.dependencies),
-
-              spec: {
-                ...omit(task.spec, ["name", "dependencies", "disabled"]),
-              },
-            })
-          }
-
-          for (const test of tests) {
-            actions.push({
-              kind: "Test",
-              type: "test",
-              name: module.name + "-" + test.name,
-              ...params.baseFields,
-
-              disabled: test.disabled,
-              build: buildAction ? buildAction.name : undefined,
-              dependencies: convertRuntimeDependencies(test.spec.dependencies),
-
-              spec: {
-                ...omit(test.spec, ["name", "dependencies", "disabled"]),
-              },
-            })
-          }
-
-          return {
-            group: {
-              // This is an annoying TypeScript limitation :P
-              kind: <"Group">"Group",
-              name: module.name,
-              path: module.path,
-              actions,
-              variables: module.variables,
-              varfiles: module.varfile ? [module.varfile] : undefined,
-            },
-          }
-        },
-
-        getModuleOutputs: async (params) => {
-          validateParams(params, moduleActionDescriptions.getModuleOutputs.paramsSchema)
-          return { outputs: { foo: "bar" } }
-        },
-
-        suggestModules: async () => {
-          return { suggestions: [] }
-        },
-      },
-    },
-  ],
-  createActionTypes: {
-    Build: [
-      {
-        name: "test",
-        docs: "Test Build action",
-        schema: joi.object(),
-        handlers: {
-          getStatus: async (_params) => {
-            return { state: "ready", detail: {}, outputs: { foo: "bar" } }
-          },
-
-          build: async (_params) => {
-            return { state: "ready", detail: {}, outputs: { foo: "bar" } }
-          },
-
-          publish: async (_params) => {
-            return { state: "ready", detail: null, outputs: {} }
-          },
-
-          run: async (params) => {
-            return {
-              moduleName: params.action.name,
-              command: params.args,
-              completedAt: now,
-              log: "bla bla",
-              success: true,
-              startedAt: now,
-              version: params.action.versionString(),
-            }
-          },
-        },
-      },
-    ],
-    Deploy: [
-      {
-        name: "test",
-        docs: "Test Deploy action",
-        schema: joi.object(),
-        outputs: outputsCfg,
-        handlers: {
-          getStatus: async (params) => {
-            return {
-              state: "ready",
-              detail: { state: "ready", detail: {} },
-              outputs: getTestPluginOutputs(params),
-            }
-          },
-
-          deploy: async (params) => {
-            // validateParams(params, moduleActionDescriptions.deployService.paramsSchema)
-            return {
-              state: "ready",
-              detail: { state: "ready", detail: {} },
-              outputs: getTestPluginOutputs(params),
-            }
-          },
-
-          delete: async (_params) => {
-            return { state: "ready", detail: { state: "ready", detail: {} }, outputs: {} }
-          },
-
-          exec: async (_params) => {
-            return {
-              code: 0,
-              output: "bla bla",
-            }
-          },
-
-          getLogs: async (_params) => {
-            return {}
-          },
-
-          run: async (params) => {
-            return {
-              moduleName: params.action.name,
-              command: ["foo"],
-              completedAt: now,
-              log: "bla bla",
-              success: true,
-              startedAt: now,
-              version: params.action.versionString(),
-            }
-          },
-
-          getPortForward: async (params) => {
-            validateParams(params, moduleActionDescriptions.getPortForward.paramsSchema)
-            return {
-              hostname: "bla",
-              port: 123,
-            }
-          },
-
-          stopPortForward: async (params) => {
-            validateParams(params, moduleActionDescriptions.stopPortForward.paramsSchema)
-            return {}
-          },
-        },
-      },
-    ],
-    Run: [
-      {
-        name: "test",
-        docs: "Test Run action",
-        schema: joi.object(),
-        outputs: outputsCfg,
-        handlers: {
-          getResult: async (params) => {
-            return {
-              state: "ready",
-              detail: {
-                moduleName: params.action.name,
-                taskName: params.action.name,
-                command: ["foo"],
-                completedAt: now,
-                log: "bla bla",
-                success: true,
-                startedAt: now,
-                version: params.action.versionString(),
-              },
-              outputs: getTestPluginOutputs(params),
-            }
-          },
-
-          run: async (params) => {
-            // Create artifacts, to test artifact copying
-            for (const artifact of params.action.getSpec().artifacts || []) {
-              await ensureFile(join(params.artifactsPath, artifact.source))
-            }
-
-            return {
-              state: "ready",
-              detail: {
-                moduleName: params.action.name,
-                taskName: params.action.name,
-                command: ["foo"],
-                completedAt: now,
-                log: "bla bla",
-                success: true,
-                startedAt: now,
-                version: params.action.versionString(),
-              },
-              outputs: getTestPluginOutputs(params),
-            }
-          },
-        },
-      },
-    ],
-    Test: [
-      {
-        name: "test",
-        docs: "Test Test action",
-        schema: joi.object(),
-        handlers: {
-          run: async (params) => {
-            // Create artifacts, to test artifact copying
-            for (const artifact of params.action.getSpec().artifacts || []) {
-              await ensureFile(join(params.artifactsPath, artifact.source))
-            }
-
-            return {
-              state: "ready",
-              detail: {
-                moduleName: params.action.name,
-                command: [],
-                completedAt: now,
-                log: "bla bla",
-                outputs: {
-                  log: "bla bla",
-                },
-                success: true,
-                startedAt: now,
-                testName: params.action.name,
-                version: params.action.versionString(),
-              },
-              outputs: getTestPluginOutputs(params),
-            }
-          },
-
-          getResult: async (params) => {
-            return {
-              state: "ready",
-              detail: {
-                moduleName: params.action.name,
-                command: [],
-                completedAt: now,
-                log: "bla bla",
-                outputs: {
-                  log: "bla bla",
-                },
-                success: true,
-                startedAt: now,
-                testName: params.action.name,
-                version: params.action.versionString(),
-              },
-              outputs: getTestPluginOutputs(params),
-            }
-          },
-        },
-      },
-    ],
-  },
-})
-
-const testPluginB = createGardenPlugin({
-  ...omit(testPlugin, ["createModuleTypes", "createActionTypes"]),
-  name: "test-plugin-b",
-})
-
-function getTestPluginOutputs(params: any) {
-  return { base: "ok", foo: params.action._config[returnWrongOutputsCfgKey] ? 123 : "ok" }
-}
-
-function validateParams(params: any, schema: CustomObjectSchema) {
-  validateSchema(
-    params,
-    schema.keys({
-      graph: joi.object(),
-    })
-  )
-}
